Factor out image diff code
